# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Demo Locker API",
    "version": "1.0.0",
    "description": "API for Demo Locker, a private-by-default demo/playlist sharing service. Most routes require a Bearer session token; media elements (<audio>/<img>) that cannot send an Authorization header instead accept the same token via a `?token=` query parameter (session token OR a per-playlist share/invite token — both are treated identically). Gated legacy routes never distinguish 'exists but private' from 'does not exist': any access failure returns a generic `{\"error\":\"not found\"}` 404 so private playlist/track/comment IDs cannot be enumerated by probing.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment (Worker or self-hosted)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token returned by POST /auth/signup or POST /auth/login, sent as `Authorization: Bearer <token>`."
      },
      "tokenQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Session token OR a share/invite token, sent as a `?token=` query parameter. Used by media elements (<audio src>, <img src>) that cannot set an Authorization header. Accepted anywhere bearerAuth is accepted on the legacy (non-/public/v1) gated routes."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "accent": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "#fc0",
              "#f80",
              "#4af",
              "#3f6",
              "#f6a",
              "#a6f",
              "#6cf",
              null
            ],
            "description": "Owner accent colour, one of the shipped palette. Null when never set — clients fall back to #fc0."
          },
          "lockerOwnerId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Null when this account owns its locker. Set to that owner's user id when this account is a collaborator on that owner's locker instead."
          }
        },
        "required": [
          "id",
          "email",
          "accent",
          "lockerOwnerId"
        ]
      },
      "Playlist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdByMe": {
            "type": "boolean",
            "description": "Attribution, NOT permission: whether the requesting session created this playlist. Computed per request. The raw creator id (playlists.created_by) is never serialized to any client on any route — a locker may hold several collaborators, and none of them may learn another's user id; the same rule as tracks.uploaded_by. Do not build a delete control from this field alone: DELETE /playlists/{id} lets the locker owner delete any playlist, so the owner gets false on every collaborator-created playlist they are nonetheless allowed to delete. The client-side rule is createdByMe OR locker ownership. The server never reads this field back — authorisation is decided from stored columns on every request. False is ambiguous and must not be rendered as \"someone else created this\": it also means no creator is recorded (rows predating the column), or that the requester has no identity to compare, which includes an anonymous share/invite holder."
          },
          "createdByName": {
            "type": [
              "string",
              "null"
            ],
            "description": "Attribution by NAME: who created this playlist. Resolved server-side from the creator's users.display_name — the label the locker owner typed when minting that person's invite — falling back to their email address when they have none (the locker owner has no invite, and so has none). The raw creator id is still never serialized: this is a name, not the id under another key. createdByMe alone could not answer this — it says \"mine / not mine\", which in a locker holding two collaborators never identifies WHICH of them. Null when no creator is recorded (rows predating the column, or a creator since removed), and null for every reader with no locker session: an anonymous share or invite holder is outside the locker and is served no names at all. Present on the caller's own rows too — rendering those as \"you\" is the client's decision. This is ATTRIBUTION, NOT PERMISSION: it says whose work the row is, and grants and gates nothing. Every control on the row behaves exactly as it did before this field existed, and the server never reads it back — authorisation is decided from stored columns on every request. Null means there is nothing to attribute and MUST render as nothing at all, never as \"unknown\" or \"null\"."
          },
          "name": {
            "type": "string"
          },
          "artworkKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "isPublic": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ownerId",
          "createdByMe",
          "name",
          "isPublic",
          "createdAt",
          "updatedAt",
          "createdByName"
        ],
        "description": "A playlist as returned by every route that serves one. created_by is never serialized: it is replaced by the per-request createdByMe boolean, so the shape is identical for a locker session and for an anonymous share/invite holder (who simply always gets false). One serializer, not two — see lib/public-playlist.ts."
      },
...
    }
  },
  "paths": {
...
    "/auth/signup": {
      "post": {
        "summary": "Create an account",
        "operationId": "authSignup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "inviteToken": {
                    "type": "string",
                    "description": "A collaborator invite token, from POST /collab/invites. An invite is its own authorisation to create an account, so a valid unredeemed one creates the account even where registration is otherwise closed, and binds it to the inviting owner's locker. Each invite is good for exactly one account: unknown, already-redeemed and expired tokens all fail with 403 and never fall through to ordinary signup."
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created, session issued.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "token": {
                      "type": "string",
                      "description": "Bearer session token, valid 30 days."
                    }
                  },
                  "required": [
                    "user",
                    "token"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing email/password, or password under 8 characters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Registration is closed, or the supplied inviteToken is not valid. Without an invite, signup is permitted only while the instance has no accounts, unless ALLOW_SIGNUP=true. With one, unknown, already-redeemed and expired tokens are refused identically, and are checked before the duplicate-email 409 so signup cannot be used to confirm an address. A valid invite is also refused when the locker has no seat left, since MAX_COLLABORATORS is re-checked at redemption and may have been lowered since the invite was minted; that refusal leaves the invite unspent and redeemable once room is made, and like every other invite refusal it is answered before the duplicate-email 409, so a full locker cannot be used to confirm an address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered. Also returned when two concurrent signups contend for one address and this one loses at the insert; if it carried an invite, that invite is handed back unspent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. See Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in",
        "operationId": "authLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session issued.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        }
                      },
                      "required": [
                        "id",
                        "email"
                      ]
                    },
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "user",
                    "token"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing email/password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. See Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get the authenticated user",
        "operationId": "authMe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": [
                    "user"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out",
        "description": "Deletes the session matching the Bearer token, if any.",
        "operationId": "authLogout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/shares/invite/{token}": {
      "get": {
        "summary": "Resolve an invite/share token",
        "description": "Unauthenticated. Unlike the other gated routes, invalid/expired tokens return distinct error messages (this endpoint is the entry point for invite links, not a resource lookup that must resist enumeration).",
        "operationId": "resolveInvite",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist and tracks the invite grants access to.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "permission": {
                      "type": "string",
                      "enum": [
                        "listen",
                        "edit"
                      ]
                    },
                    "playlist": {
                      "$ref": "#/components/schemas/Playlist"
                    },
                    "tracks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Track"
                      }
                    },
                    "accent": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "#fc0",
                        "#f80",
                        "#4af",
                        "#3f6",
                        "#f6a",
                        "#a6f",
                        "#6cf",
                        null
                      ],
                      "description": "The playlist owner's accent colour, so a shared locker keeps its colours for listeners. Null when the owner never set one."
                    }
                  },
                  "required": [
                    "permission",
                    "playlist",
                    "tracks",
                    "accent"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Token does not exist, or the playlist it referenced was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Token exists but is past its expiresAt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/auth/accent": {
      "post": {
        "summary": "Set the account accent colour",
        "description": "Account-level, not per-browser: this is the colour listeners see on a share link. Values outside the shipped palette are rejected — the value is written into a CSS custom property client-side.",
        "operationId": "setAccent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accent": {
                    "type": "string",
                    "enum": [
                      "#fc0",
                      "#f80",
                      "#4af",
                      "#3f6",
                      "#f6a",
                      "#a6f",
                      "#6cf"
                    ]
                  }
                },
                "required": [
                  "accent"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accent stored.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accent": {
                      "type": "string",
                      "enum": [
                        "#fc0",
                        "#f80",
                        "#4af",
                        "#3f6",
                        "#f6a",
                        "#a6f",
                        "#6cf"
                      ]
                    }
                  },
                  "required": [
                    "accent"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Accent outside the supported palette.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}